+++ projects/example/src/app/app.routes.ts
@@ -37,6 +37,8 @@ export const routes: Routes = [
   { path: 'pk-form', loadComponent: () => import('./pages/pk-form/pk-form-page').then(m => m.PkFormPage) },
   { path: 'pk-layout', loadComponent: () => import('./pages/pk-layout/pk-layout-page').then(m => m.PkLayoutPage) },
   { path: 'pk-markdown-viewer', loadComponent: () => import('./pages/pk-markdown-viewer/pk-markdown-viewer-page').then(m => m.PkMarkdownViewerPage) },
+  { path: 'pk-barcode', loadComponent: () => import('./pages/pk-barcode/pk-barcode-page').then(m => m.PkBarcodePage) },
+  { path: 'pk-qrcode', loadComponent: () => import('./pages/pk-qrcode/pk-qrcode-page').then(m => m.PkQrcodePage) },
   // Examples
   { path: 'examples/login', loadComponent: () => import('./pages/examples/login/login-example').then(m => m.LoginExample) },
   { path: 'examples/chat', loadComponent: () => import('./pages/examples/chat/chat-example').then(m => m.ChatExample) },