+++ projects/ngx-pk-ui/src/lib/pk-qrcode/pk-qrcode.model.ts
@@ -1,0 +1,4 @@
+export type PkQrEcLevel = 'L' | 'M' | 'Q' | 'H';
+
+/** @internal */
+export type PkQrMode = 'numeric' | 'alphanumeric' | 'byte';